[#94] cors 설정 파일분리
