+++ app/src/main/kotlin/com/santaclose/app/AppApplication.kt
@@ -4,40 +4,13 @@ import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.autoconfigure.domain.EntityScan
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan
 import org.springframework.boot.runApplication
-import org.springframework.context.annotation.Bean
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing
-import org.springframework.http.HttpMethod
-import org.springframework.web.cors.CorsConfiguration
-import org.springframework.web.cors.reactive.CorsWebFilter
-import org.springframework.web.cors.reactive.UrlBasedCorsConfigurationSource
 
 @EnableJpaAuditing
 @ConfigurationPropertiesScan
 @EntityScan(basePackages = ["com.santaclose.lib.entity"])
 @SpringBootApplication
-class AppApplication {
-//    @Bean
-//    fun wiringFactory() =
-//        KotlinDirectiveWiringFactory(manualWiring = mapOf("auth" to AuthSchemaDirectiveWiring()))
-//
-//    @Bean
-//    fun hooks(wiringFactory: KotlinDirectiveWiringFactory) = CustomSchemaGeneratorHooks(wiringFactory)
-
-    @Bean
-    fun corsFilter(): CorsWebFilter {
-        val config =
-            CorsConfiguration().apply {
-                allowCredentials = true
-                addAllowedOrigin("https://studio.apollographql.com")
-                addAllowedHeader("*")
-                addAllowedMethod(HttpMethod.POST.name)
-            }
-
-        return UrlBasedCorsConfigurationSource()
-            .apply { registerCorsConfiguration("/graphql/**", config) }
-            .let(::CorsWebFilter)
-    }
-}
+class AppApplication
 
 fun main(args: Array<String>) {
     runApplication<AppApplication>(*args)
